requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
Django==2.2.2
django-extra-views==0.12.0
django-pytest==0.2.0
django-webtest==1.9.5
importlib-metadata==0.17
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.2
pytest-django==3.5.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
waitress==1.3.0
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
zipp==0.5.1
